check for existent of user  before retrieve data

+++ AppLogic/UserManager.cs
@@ -3,7 +3,9 @@ using DataAccess.Mappers;
 using DTO.Users;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Data;
+using System.Diagnostics.CodeAnalysis;
 using System.IO.MemoryMappedFiles;
 using System.Linq;
 using System.Text;
@@ -18,8 +20,9 @@ namespace AppLogic
 
         public void CreateUser(UserInput user)
         {
-            DbUser userFound = mapper.GetByEmail(user.email);
-            if (userFound.email != null)
+            var userExist = ExistUser(user.email);
+            
+            if (userExist)
             {
                 throw new Exception("The email is already registered");
             }
@@ -29,8 +32,12 @@ namespace AppLogic
             _ = mapper.Create(user);
         }
 
-        public DbUser GetUserById(int id) { 
-         return mapper.GetById(id);
+        public DbUser GetUserById(int id) {
+            var existUser = ExistUser(id);
+            if (!existUser){
+                throw new Exception("User not found");
+            }   
+            return mapper.GetById(id);
         }
 
         public List<DbUser> GetUsers()
@@ -40,23 +47,38 @@ namespace AppLogic
 
         public  DbUser GetUserByEmail(string email)
         {
+           var existUser = ExistUser(email);
+           if (!existUser){
+            throw new Exception("User not found");
+           }
             return mapper.GetByEmail(email);
         }
 
         public void UpdateUser(DbUser user)
         {
-            
+            var existUser = ExistUser(user.id);
+            if (!existUser){
+                throw new Exception("User doesn't exist");
+            } 
             _ = mapper.Update(user);
         }
 
         public void DeleteUser(int id)
         {
+            var existUser = ExistUser(id);
+            if (!existUser){
+                throw new Exception("No user to delete");
+            }
             _ = mapper.Delete(id);
 
         }
 
         public string Login(string email, string password)
         {
+            var existUser = ExistUser(email);
+            if (!existUser){
+                throw new Exception("Invalid credentials");
+            }
             var result  = mapper.Login(email, password);
             return result;
         }
@@ -89,15 +111,32 @@ namespace AppLogic
             }
         }
 
-        public DbUser CastUserInput(UserInput userInput)
-        {
-           DbUser user = new DbUser();
-            user.name = userInput.name;
-            user.lastName = userInput.lastName;
-            user.email = userInput.email;
-            user.birthDay = userInput.birthDay;
-            user.active = null;
-            return user;    
+        // public DbUser CastUserInput(UserInput userInput)
+        // {
+        //    DbUser user = new DbUser();
+        //     user.name = userInput.name;
+        //     user.lastName = userInput.lastName;
+        //     user.email = userInput.email;
+        //     user.birthDay = userInput.birthDay;
+        //     user.active = null;
+        //     return user;    
+        // }
+
+        public bool ExistUser(int id){
+            var userFound = mapper.GetById(id);
+            if (userFound.id == 0){
+                return false;
+            }
+            return true;
+            
+
+        }
+        public bool ExistUser(string email){
+            var userFound = mapper.GetByEmail(email);
+            if (userFound.id == 0){
+                return false;
+            }
+            return true;
         }
     }
 }
